refactored a bunch of shit
