+++ logic/logic.go
@@ -1,0 +1,9 @@
+package logic
+
+import "runtime"
+
+func GetMemoryUsage() uint64 {
+	var m runtime.MemStats
+	runtime.ReadMemStats(&m)
+	return m.Alloc
+}
