+++ routes/v1/v1.go
@@ -15,7 +15,9 @@ func New(db database.Database, app *fiber.App) V1 {
 	}
 
 	app.Get("/ping", v1.Ping)
-	app.Get("/users", v1.GetUsersAll)
+	app.Get("/status", v1.Status)
+
+	app.Get("/users/all", v1.GetUsersAll)
 	app.Get("/users/uuid/:uuid", v1.GetUserByUuid)
 	app.Get("/users/email/:email", v1.GetUserByEmail)
 	app.Get("/users/username/:username", v1.GetUserByUsername)
