+++ types/types.go
@@ -1,0 +1,5 @@
+package types
+
+type BackendStatus struct {
+	MemoryUsage uint64 `json:"memory_usage" xml:"memory_usage" form:"memory_usage"`
+}
