+++ version.txt
@@ -1,1 +1,1 @@
-1.3.0
+3.0.1
